scraper/requirements.txt
# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
scrapy==2.11.2
selenium==4.21.0
webdriver-manager==4.0.1
playwright==1.52.0

# AI-powered fallback strategies
firecrawl-py==1.5.0
anthropic>=0.40.0

# Data processing
python-dateutil==2.9.0

# Email notifications
sendgrid==6.11.0

# Linting / formatting (dev)
ruff==0.4.9

# Testing
pytest==8.2.2
pytest-cov==5.0.0
responses==0.25.3

# Utilities
python-dotenv==1.0.1
